Rearrange mirai code implementations
